DirectSound streaming now working

+++ jukebox_test/demo/main.cpp
@@ -24,11 +24,11 @@
 
 std::string formatDuration(double duration) {
 	std::cout << "*** " << duration << std::endl;
-	size_t hr = duration/3600;
+	int hr = duration/3600;
 	duration -= hr*3600;
-	size_t min = duration/60;
+	int min = duration/60;
 	duration -= min*60;
-	size_t secs = duration;
+	int secs = duration;
 
 	return
 		std::to_string(hr) + ":" +
